python/agents/sft-runner-starter-pack/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sft-runner-starter-pack"
version = "0.1.0"
description = "Supervised Fine Tuning Starter Pack, using Gemini and ADK."
authors = [
  { name = "Yogesh Kothari", email = "yogeshkothari@google.com" },
  { name = "Nishant Welpulwar", email = "nishantw@google.com" },
  { name = "Suddhasatwa Bhaumik", email = "suddhasatwa@google.com" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
dependencies = [
    "python-dotenv",
    "python-multipart",
    "google-cloud-aiplatform",
    "google-generativeai",
    "google-adk",
    "google-cloud-storage",
    "db-dtypes",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "vertexai",
    "fsspec",
    "gcsfs",
    "aiohttp",
    "click",
    "grpcio",
    "google-cloud-aiplatform[adk,agent_engines]",
    "google-cloud-aiplatform[evaluation]",
    "pandas",
    "pydantic",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "streamlit",
    "sseclient",
    "sqlalchemy",
]
